The python code and data for SNA
Similar_names (3) contains the Python code needed to convert raw patent data into the files containing GephiE in their name. 
SNA FC + FCV TIS R1 is used as inputs for the later parts of Simlar_names.
The raw patent data is not included as the file sizes are too large.